requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
soundfile>=0.12.1 